requirements.txt
altair==4.2.2
attrs==23.1.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
entrypoints==0.4
fonttools==4.47.0
gitdb==4.0.11
GitPython==3.1.40
idna==3.6
importlib-metadata==7.0.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lazypredict==0.2.12
lightgbm==4.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mdurl==0.1.2
numpy==1.26.2
packaging==23.2
pandas==2.1.2
patsy==0.5.5
Pillow==10.1.0
protobuf==3.20.3
pyarrow==14.0.2
pydeck==0.8.1b0
Pygments==2.17.2
Pympler==1.0.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
referencing==0.32.0
requests==2.31.0
rich==13.7.0
rpds-py==0.16.2
scikit-learn==1.3.0
scipy==1.11.4
six==1.16.0
smmap==5.0.1
statsmodels==0.14.1
streamlit==1.22.0
tenacity==8.2.3
threadpoolctl==3.2.0
toml==0.10.2
toolz==0.12.0
tornado==6.4
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
watchdog==3.0.0
xgboost==2.0.3
yellowbrick==1.5
zipp==3.17.0
